move test to selection
